Fixed some things and implemented checking and completing own tasks

+++ src/main/resources/static/js/tasks.js
@@ -1,3 +1,4 @@
+document.getElementById("load-own-tasks").addEventListener("click", loadOwnTasks);
 document.getElementById("load-tasks").addEventListener("click", loadTasks);
 document.getElementById("add-task").addEventListener("click", createTask);
 document.getElementById("logout").addEventListener("click", logout);
@@ -11,9 +12,64 @@ if (!token) {
 	window.location.href = "/login.html"
 }
 
+const loggedAsAdmin = localStorage.getItem("isAdmin") === "true";
+
+if (!loggedAsAdmin) {
+	const adminActions = document.getElementById("adminActions");
+	if (adminActions) {
+		adminActions.style.display = "none";
+	}
+} else {
+	const title = document.getElementById("title");
+	title.innerHTML += " (Logged in as admin)";
+}
+
 function logout() {
 	document.cookie = "jwt=; Path=/;";
+	localStorage.removeItem("isAdmin");
+	localStorage.removeItem("username");
 	window.location.href = "/login.html"
+}
+
+function loadOwnTasks() {
+	fetch("/tasks/own", {
+		headers: { "Authorization": "Bearer " + token }
+	})
+		.then(res => res.json())
+		.then(tasks => {
+			const list = document.getElementById("own-tasks-list");
+            list.innerHTML = "";
+
+            tasks.forEach(t => {
+                const li = document.createElement("li");
+				let completion = ""
+				if (t.completed) {
+					completion = `
+						Completed! 
+						<button onClick="changeCompletionTask(${t.taskId})">Mark as not yet completed</button>
+					`;
+				} else {
+					completion = `
+						Not yet completed 
+						<button onClick="changeCompletionTask(${t.taskId})">Mark as completed</button>
+					`;
+				}
+                li.innerHTML = `
+                    Assignment: ${t.assignment}<br>
+                    ${completion}
+                    <hr>
+                `;
+                list.appendChild(li);
+            });
+		});
+}
+
+function changeCompletionTask(taskId) {
+	fetch(`/tasks/complete/${taskId}`, {
+		method: "PUT",
+		headers: { "Content-Type": "application/json", "Authorization": "Bearer " + token },
+	})
+		.then(() => loadOwnTasks());
 }
 
 function loadTasks() {
@@ -27,13 +83,20 @@ function loadTasks() {
             list.innerHTML = "";
 
             tasks.forEach(t => {
+				let adminButtons = "";
+				if (loggedAsAdmin) {
+					adminButtons = `
+						<button onclick="deleteTask(${t.taskId})">❌</button>
+						<button onclick="updateTask(${t.taskId})">✏️</button>
+					`;
+				}
+				
                 const li = document.createElement("li");
                 li.innerHTML = `
                     ID: ${t.taskId} — User: ${t.userId}<br>
                     Assigned: ${t.assigned} | Completed: ${t.completed}<br>
                     Assignment: ${t.assignment}<br>
-                    <button onclick="deleteTask(${t.taskId})">❌</button>
-                    <button onclick="updateTask(${t.taskId})">✏️</button>
+                    ${adminButtons}
                     <hr>
                 `;
                 list.appendChild(li);
@@ -48,7 +111,7 @@ function createTask() {
     const assigned = document.getElementById("task-assigned").checked;
 
     fetch("/tasks", {
-        method: "POST",
+        method: "PUT",
         headers: { "Content-Type": "application/json", "Authorization": "Bearer " + token },
         body: JSON.stringify({ userId, assignment, completed, assigned })
     })
@@ -65,7 +128,7 @@ function updateTask(taskId) {
     const completed = document.getElementById("task-completed").checked;
     const assigned = document.getElementById("task-assigned").checked;
 
-    fetch(`/tasks/${taskId}`, {
+    fetch(`/tasks/update/${taskId}`, {
         method: "PUT",
         headers: { "Content-Type": "application/json", "Authorization": "Bearer " + token },
         body: JSON.stringify({ userId, assignment, completed, assigned })
